Move all implementation from BreakIterator into RuleBasedBreakIterator
